object_detection/requirements.txt
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.4.0
setuptools==57.4.0
torch==1.10.0+cu102
torchvision==0.11.1+cu102
